+++ src/sll/include/sll/node.h
@@ -160,11 +160,11 @@ __SLL_EXTERNAL void sll_optimize_node(sll_compilation_data_t* c_dt,sll_internal_
  * \desc Docs!
  * \arg sll_compilation_data_t* c_dt
  * \arg sll_internal_function_table_t* i_ft
- * \arg sll_import_loader_t il
+ * \arg sll_import_resolver_t il
  * \arg sll_error_t* e
  * \ret sll_bool_t
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_parse_all_nodes(sll_compilation_data_t* c_dt,sll_internal_function_table_t* i_ft,sll_import_loader_t il,sll_error_t* e);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_parse_all_nodes(sll_compilation_data_t* c_dt,sll_internal_function_table_t* i_ft,sll_import_resolver_t il,sll_error_t* e);
 
 
 
@@ -175,12 +175,12 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_parse_all_nodes(sll_compilation
  * \desc Docs!
  * \arg sll_compilation_data_t* c_dt
  * \arg sll_internal_function_table_t* i_ft
- * \arg sll_import_loader_t il
+ * \arg sll_import_resolver_t il
  * \arg sll_error_t* e
  * \arg sll_node_t** o
  * \ret sll_bool_t
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_parse_node(sll_compilation_data_t* c_dt,sll_internal_function_table_t* i_ft,sll_import_loader_t il,sll_error_t* e,sll_node_t** o);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_parse_node(sll_compilation_data_t* c_dt,sll_internal_function_table_t* i_ft,sll_import_resolver_t il,sll_error_t* e,sll_node_t** o);