+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - Implemented `sll_api_base64_decode` and `base64$decode` ([#169] and [#172])
 - Implemented `sll_api_base64_encode` and `base64$encode` ([#169] and [#172])
 - Implemented `sll_api_json_stringify` and `json$stringify` ([#58])
+- Version tag macro constant: `SLL_VERSION_TAG`
+
+### Changed
+
+- Renamed `sll_import_loader_t` to `sll_import_resolver_t`
 
 ## [0.6.34] - 2021-12-30